Parry system and spawner
Created parry system. Use right click to parry. Getting a perfect parry boost the speed and damage more. You can set all multipliers in the editor. You can also change how long the shield is on the screen in the editor. You can also try adjusting the enemy projectile screen in the editor.

There is now an object that spawns enemies. The key to activate it is bound to "R" right now but this will obviously not be in the main game.

To make it so they don't follow you around just uncheck the movement box in the editor in the enemy prefab in the prefab folder.

All of these changes are a work in progress and subject to change in the future.